reorganized a bit
